+++ docs/source/changes.md
@@ -11,7 +11,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 - {pull}`517` raises an error when the configuration file contains a non-existing path
   (fixes #514). It also warns if the path is configured as a string and not a list of
   strings.
-- {pull}`518` raises an error when builtin functions are wrapped with
+- {pull}`519` raises an error when builtin functions are wrapped with
   {func}`~pytask.task`. Closes {issue}`512`.
 
 ## 0.4.4 - 2023-12-04